refactor: move Content-Disposition generation to utils package
- Replaced the inline Content-Disposition generation in the Doubao and DoubaoShare drivers with a call to the new utils.GenerateContentDisposition function.
- Removed the redundant generateContentDisposition functions from both drivers.
- Updated related files to utilize the new utility function for consistent Content-Disposition header formatting.
