+++ server/s3/backend.go
@@ -121,8 +121,9 @@ func (b *s3Backend) HeadObject(ctx context.Context, bucketName, objectName strin
 	// hash := getFileHashByte(fobj)
 
 	meta := map[string]string{
-		"Last-Modified": node.ModTime().Format(timeFormat),
-		"Content-Type":  utils.GetMimeType(fp),
+		"Last-Modified":      node.ModTime().Format(timeFormat),
+		"Content-Disposition": utils.GenerateContentDisposition(path.Base(objectName)),
+		"Content-Type":       utils.GetMimeType(fp),
 	}
 
 	if val, ok := b.meta.Load(fp); ok {
@@ -213,8 +214,9 @@ func (b *s3Backend) GetObject(ctx context.Context, bucketName, objectName string
 	}
 
 	meta := map[string]string{
-		"Last-Modified": node.ModTime().Format(timeFormat),
-		"Content-Type":  utils.GetMimeType(fp),
+		"Last-Modified":      node.ModTime().Format(timeFormat),
+		"Content-Disposition": utils.GenerateContentDisposition(file.GetName()),
+		"Content-Type":       utils.GetMimeType(fp),
 	}
 
 	if val, ok := b.meta.Load(fp); ok {
@@ -328,7 +330,7 @@ func (b *s3Backend) PutObject(
 func (b *s3Backend) DeleteMulti(ctx context.Context, bucketName string, objects ...string) (result gofakes3.MultiDeleteResult, rerr error) {
 	for _, object := range objects {
 		if err := b.deleteObject(ctx, bucketName, object); err != nil {
-			utils.Log.Errorf("serve s3", "delete object failed: %v", err)
+			log.Errorf("delete object failed: %v", err)
 			result.Error = append(result.Error, gofakes3.ErrorResult{
 				Code:    gofakes3.ErrInternal,
 				Message: gofakes3.ErrInternal.Message(),
